+++ src/ssr.ts
@@ -13,7 +13,7 @@ import { getComponentCss } from "./getComponentCss.ts";
  *
  * @example
  * ```tsx
- * import { byoHtml, HtmlShell } from "@fullsoak/fullsoak";
+ * import { byoHtml, HtmlShell } from "@fullsoak/fullsoak/batteries";
  * byoHtml(
  *   <HtmlShell
  *     componentName="MyApp"